Submit solutions to the Repository
The solutions are saved in Automation folders, readme file included.
Also the Web_API_Demo application is included for testing purpose if need be.

I found one issue in API Testing when adding Item without ID. From MSTest, it's an error but not from Postman. Added one Performance Test and we can extend to cover all endpoints.

Others are good.